// src/graphql/mutation/__generated__/BuyItem.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { ProductType } from "./../../../../__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: BuyItem
// ====================================================

export interface BuyItem_buyItem {
  __typename: "Transaction";
  id: string;
  productName: string;
  productType: ProductType;
  itemValue: string;
  itemPrice: number;
  primaryFormName: string | null;
  secondaryFormName: string | null;
  primaryFormValue: string | null;
  secondaryFormValue: string | null;
  proofOfPayment: string | null;
  voucher: string | null;
  message: string | null;
  pending: boolean;
  success: boolean;
}

export interface BuyItem {
  buyItem: BuyItem_buyItem;
}

export interface BuyItemVariables {
  id: string;
  primaryFormValue?: string | null;
  secondaryFormValue?: string | null;
}